requirements.txt
# Web框架
fastapi>=0.115,<0.116
uvicorn>=0.34.2,<0.35.0
python-multipart>=0.0.20,<0.0.21
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
email-validator==2.1.0.post1

# 数据库
sqlalchemy==2.0.27
alembic==1.13.1
aiosqlite==0.20.0

# 数据处理
pandas==2.2.0
numpy==1.26.4
akshare==1.16.95
openpyxl>=3.1.5,<4.0.0

# 工具
python-dotenv==1.0.1
pydantic==2.6.1
pydantic-settings==2.1.0
requests>=2.32.1,<3.0.0
aiohttp==3.11.13
tenacity==8.2.3

# 开发工具
black==24.1.1
isort==5.13.2
flake8==7.0.0
pytest==8.0.0
pytest-asyncio==0.23.5
httpx==0.26.0

PyJWT>=2.0.0 